requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Brotli==1.0.9
dj-database-url==1.0.0
Django==4.1.1
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
whitenoise==6.2.0
